Changed Structure of backend and frontend (also added a sample test case)
