+++ Backend/Tests/test_balance_search.py
@@ -1,0 +1,3 @@
+import sys
+import os
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
